+++ examples/example_reinforcement_patterns.py
@@ -3,7 +3,7 @@ Exemple : Différentes dispositions de ferraillage
 """
 
 import numpy as np
-import opensection as ps
+import opensection as ops
 
 try:
     import matplotlib.pyplot as plt